Change density return type from int to double

+++ SparseMatrix.h
@@ -25,7 +25,7 @@ public:
     int get(int xPos, int yPos);
     void remove(int xPos, int yPos);
     void printStoredValues();
-    int density();  
+    double density();  
     SparseMatrix* multiply(SparseMatrix* second);
 
     int getElement(int xPos, int yPos);
